# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cheeta Api Docs",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://192.168.100.20/cheeta/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "tag/": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "tags",
                "description": "get tags",
                "operationId": "51e5933cdc0520c3ae9d160c6e83b722",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "09388985617"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/order/edit/{{barcode}}": {
            "post": {
                "tags": [
                    "order"
                ],
                "summary": "ویرایش سفارش",
                "description": "edit order",
                "operationId": "updateorder",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "barcode": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "09388985617"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/order/register": {
            "post": {
                "tags": [
                    "order"
                ],
                "summary": "ثبت سفارش",
                "description": "User Register here",
                "operationId": "addorder",
                "parameters": [
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "description": "api token",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text"
                                    },
                                    "email": {
                                        "type": "text"
                                    },
                                    "password": {
                                        "type": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "persistent": {
                                        "type": "boolean",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "oauth2": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/login": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Sign in",
                "description": "Login by PHONE",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "09388985617"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "security": [
        {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            },
            "passport": []
        }
    ]
}